Fix int in models

+++ Models/ExerciseStats.cs
@@ -5,7 +5,7 @@
         public string ExerciseName { get; set; }
         public int Count { get; set; }
         public int TotalReps { get; set; }
-        public decimal AvgWeight { get; set; }
-        public decimal MaxWeight { get; set; }
+        public double AvgWeight { get; set; }
+        public int MaxWeight { get; set; }
     }
 }
